--- src/domain/auth/router.ts
@@ -1,10 +1,0 @@
-import express from "express"
-import * as Controller from "./controller"
-
-const authRouter = express.Router()
-
-// endpoint de creación de usuario
-authRouter.post("/register", Controller.register)
-authRouter.post("/login", Controller.login)
-
-export default authRouter
