+++ src/domain/users/repository.ts
@@ -12,6 +12,18 @@ export const getUsers = async () => {
   return users
 }
 
+export const createUser = async (user: any) => {
+  const email: string = user.email
+  const findEmail = await Users.findOneBy({ email: email })
+
+  //si no existe, lo crea y no retorna nada o undefined para mandar el response en controller
+  if (!findEmail) {
+    const newUser = await Users.create(user).save()
+    return undefined
+  } else {
+    return findEmail
+  }
+}
 
 export const getSingleUser = async (userId: number) => {
   const user = await Users.findOneBy({ id: userId })
@@ -19,4 +31,21 @@ export const getSingleUser = async (userId: number) => {
   return user
 }
 
-
+export const findByEmail = async (email: string) => {
+  const user = await Users.findOne({
+    where: {
+      email: email,
+    },
+    relations: { role_id: true },
+    select: {
+      id: true,
+      password: true,
+      email: true,
+      role_id: {
+        id: true,
+        name: true,
+      },
+    },
+  })
+  return user
+}
