+++ src/middlewheres/auth.ts
@@ -2,31 +2,33 @@ import { NextFunction, Request, Response } from "express"
 import jwt from "jsonwebtoken"
 import { TokenData } from "../types"
 
-export const auth = async (req: Request, res: Response, next: NextFunction) => {
+export const auth = (req: Request, res: Response, next: NextFunction) => {
+  const token = req.headers.authorization?.split(" ")[1]
 
+  if (!token) {
+    res.status(401).json({
+      success: false,
+      message: "UNAUTHORIZED",
+    })
+    return
+  }
 
   try {
-    const token = req.headers.authorization?.split(" ")[1]
-    console.log(token)
+    const decoded = jwt.verify(token, "secreto") as TokenData
 
-    if (!token) {
-      return res.status(401).json({
-        success: false,
-        message: "UNAUTHORIZED",
-      })
+    // Modify request object to include payload
+    req.tokenData = {
+      userId: decoded.userId,
+      userRole: decoded.userRole,
     }
-
-    const decoded = jwt.verify(token, process.env.JWT_SECRET as string)
-
-
-    req.tokenData = decoded as TokenData
 
     next()
   } catch (error) {
-    return res.status(500).json({
+    res.status(401).json({
       success: false,
-      message: "JWT NOT VALID OR MALFORMED",
-      error: error,
+      error: "Unauthorized",
+      message: "Invalid or malformed JWT token",
     })
+    return
   }
 }
