+++ src/middlewheres/isSuperAdmin.ts
@@ -1,22 +1,18 @@
-import { NextFunction, Request, Response } from "express"
+import { NextFunction, Request, Response } from "express";
 
-export const isSuperAdmin = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
-  try {
-    if (req.tokenData.roleName !== "super_admin") {
-      return res.status(401).json({
-        success: false,
-        message: "UNAUTHORIZED",
-      })
+export const isSuperAdmin = (req:Request, res: Response, next: NextFunction) =>{
+
+    try {
+        if(req.tokenData.userRole !== "super_admin"){
+           return res.status(401).json({
+                success: false,
+                error: "Unauthorized",
+              })
+        }
+
+        next()
+    } catch (error) {
+        console.log(error)
     }
-    next()
-  } catch (error) {
-    res.status(500).json({
-      success: false,
-      message: "You dont have permisions",
-    })
-  }
+
 }
