+++ src/types/index.d.ts
@@ -1,13 +1,15 @@
 export type TokenData = {
-  userId: number
-  roleName: string
-}
-
-declare global {
-  // Express
-  namespace Express {
-    export interface Request {
-      tokenData: TokenData
+    userId: number;
+    userRole: string;
+  };
+  
+  declare global {
+    // Express
+    namespace Express {
+      export interface Request {
+        tokenData: TokenData;
+      }
     }
   }
-}
+
+
